バリデーション・ログインチェックを追加

+++ src/jp/co/aforce/servlets/admin/AdminDeleteItemServlet.java
@@ -9,6 +9,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
+import jp.co.aforce.beans.RoleBean;
 import jp.co.aforce.models.ItemDAO;
 import jp.co.aforce.parameters.MessageParameter;
 
@@ -40,18 +41,24 @@ public class AdminDeleteItemServlet extends HttpServlet {
 	 */
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		HttpSession session = request.getSession();
-		String itemId = request.getParameter("deleteItem");
-		ItemDAO dao = new ItemDAO();
-		try {
-			String itemName = session.getAttribute("itemName").toString();
-			dao.deleteItemById(Integer.parseInt(itemId));
-			session.setAttribute("itemInfoBean", dao.adminSearchItem(itemName));
-			response.sendRedirect("/ShoppingSite/views/admin/adminItemSearch.jsp");
-		} catch (Exception e) {
-			session.setAttribute("adminSearchItemMessage", MessageParameter.SYSTEM_ERROR);
-			response.sendRedirect("/ShoppingSite/views/admin/adminItemSearch.jsp");
-			e.printStackTrace();
+		RoleBean bean = (RoleBean) session.getAttribute("userInfo");
+		if(bean == null || !bean.getRole().equals("admin")) {
+			response.sendRedirect("/ShoppingSite/views/login/login.jsp");
+		} else {
+			String itemId = request.getParameter("deleteItem");
+			ItemDAO dao = new ItemDAO();
+			try {
+				String itemName = session.getAttribute("itemName").toString();
+				dao.deleteItemById(Integer.parseInt(itemId));
+				session.setAttribute("itemInfoBean", dao.adminSearchItem(itemName));
+				response.sendRedirect("/ShoppingSite/views/admin/adminItemSearch.jsp");
+			} catch (Exception e) {
+				session.setAttribute("adminSearchItemMessage", MessageParameter.SYSTEM_ERROR);
+				response.sendRedirect("/ShoppingSite/views/admin/adminItemSearch.jsp");
+				e.printStackTrace();
+			}
 		}
+
 	}
 
 }
