+++ src/jp/co/aforce/servlets/consumer/DeleteCartServlet.java
@@ -9,6 +9,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
+import jp.co.aforce.beans.RoleBean;
 import jp.co.aforce.models.CartDAO;
 import jp.co.aforce.parameters.MessageParameter;
 
@@ -40,16 +41,21 @@ public class DeleteCartServlet extends HttpServlet {
 	 */
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		HttpSession session = request.getSession();
-		int itemId = Integer.parseInt(request.getParameter("deleteCart"));
-		CartDAO cartDAO = new CartDAO();
-		try {
-			cartDAO.deleteCart(itemId);
-			request.getRequestDispatcher("/myCartInfoServlet").include(request, response);
-			session.setAttribute("cartMessage", MessageParameter.DELETE_COMPLETE);
-			response.sendRedirect("/ShoppingSite/views/consumer/consumer_cart.jsp");
-		} catch (Exception e) {
-			// TODO 自動生成された catch ブロック
-			e.printStackTrace();
+		RoleBean roleBean = (RoleBean) session.getAttribute("userInfo");
+		if(roleBean == null || !roleBean.getRole().equals("consumer")) {
+			response.sendRedirect("/ShoppingSite/views/login/login.jsp");
+		} else {
+			int itemId = Integer.parseInt(request.getParameter("deleteCart"));
+			CartDAO cartDAO = new CartDAO();
+			try {
+				cartDAO.deleteCart(itemId);
+				request.getRequestDispatcher("/myCartInfoServlet").include(request, response);
+				session.setAttribute("cartMessage", MessageParameter.DELETE_COMPLETE);
+				response.sendRedirect("/ShoppingSite/views/consumer/consumer_cart.jsp");
+			} catch (Exception e) {
+				// TODO 自動生成された catch ブロック
+				e.printStackTrace();
+			}
 		}
 	}
